refactor: OfflineMemory
* Move OfflineMemory into separate module
* Delete unused Memory(Read|Write)Record
