+++ crates/vm/src/system/memory/controller/interface.rs
@@ -1,11 +1,10 @@
 use openvm_stark_backend::p3_field::PrimeField32;
 
 use crate::system::memory::{
-    controller::memory::MemoryImage,
     merkle::{DirectCompressionBus, MemoryMerkleChip},
     persistent::PersistentBoundaryChip,
     volatile::VolatileBoundaryChip,
-    CHUNK,
+    MemoryImage, CHUNK,
 };
 
 #[allow(clippy::large_enum_variant)]
