+++ crates/vm/src/system/memory/controller/mod.rs
@@ -8,7 +8,6 @@ use std::{
 };
 
 use getset::Getters;
-use memory::MemoryRecord;
 use openvm_circuit_primitives::{
     assert_less_than::{AssertLtSubAir, LessThanAuxCols},
     is_zero::IsZeroSubAir,
@@ -26,6 +25,7 @@ use openvm_stark_backend::{
     rap::AnyRap,
     Chip, ChipUsageGetter,
 };
+use rustc_hash::FxHashMap;
 use serde::{Deserialize, Serialize};
 
 use self::interface::MemoryInterface;
@@ -34,13 +34,14 @@ use crate::{
     arch::{hasher::HasherChip, MemoryConfig},
     system::memory::{
         adapter::AccessAdapterInventory,
-        controller::memory::{MemoryLogEntry, INITIAL_TIMESTAMP},
         dimensions::MemoryDimensions,
         merkle::{MemoryMerkleBus, MemoryMerkleChip},
+        offline::{MemoryRecord, OfflineMemory, INITIAL_TIMESTAMP},
         offline_checker::{
             MemoryBridge, MemoryBus, MemoryReadAuxCols, MemoryReadOrImmediateAuxCols,
             MemoryWriteAuxCols, AUX_LEN,
         },
+        online::{Address, Memory, MemoryLogEntry},
         persistent::PersistentBoundaryChip,
         tree::MemoryNode,
     },
@@ -48,16 +49,17 @@ use crate::{
 
 pub mod dimensions;
 mod interface;
-pub(super) mod memory;
-pub use memory::{MemoryImage, MemoryReadRecord, MemoryWriteRecord, OfflineMemory, RecordId};
-
-pub(crate) use crate::system::memory::controller::memory::Memory;
 
 pub const CHUNK: usize = 8;
 /// The offset of the Merkle AIR in AIRs of MemoryController.
 pub const MERKLE_AIR_OFFSET: usize = 1;
 /// The offset of the boundary AIR in AIRs of MemoryController.
 pub const BOUNDARY_AIR_OFFSET: usize = 0;
+
+#[derive(Debug, Clone, Copy)]
+pub struct RecordId(pub usize);
+
+pub type MemoryImage<F> = FxHashMap<Address, F>;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct TimestampedValues<T, const N: usize> {
@@ -691,9 +693,9 @@ impl<F: PrimeField32> MemoryController<F> {
 
 #[derive(Clone, Debug)]
 pub struct MemoryAuxColsFactory<T> {
-    range_checker: Arc<VariableRangeCheckerChip>,
-    timestamp_lt_air: AssertLtSubAir,
-    _marker: PhantomData<T>,
+    pub(crate) range_checker: Arc<VariableRangeCheckerChip>,
+    pub(crate) timestamp_lt_air: AssertLtSubAir,
+    pub(crate) _marker: PhantomData<T>,
 }
 
 // NOTE[jpw]: The `make_*_aux_cols` functions should be thread-safe so they can be used in parallelized trace generation.
