+++ crates/vm/src/system/memory/mod.rs
@@ -3,7 +3,9 @@ use openvm_circuit_primitives_derive::AlignedBorrow;
 mod adapter;
 mod controller;
 pub mod merkle;
+mod offline;
 pub mod offline_checker;
+mod online;
 mod persistent;
 #[cfg(test)]
 mod tests;
@@ -11,6 +13,7 @@ pub mod tree;
 mod volatile;
 
 pub use controller::*;
+pub use offline::*;
 
 #[derive(PartialEq, Copy, Clone, Debug, Eq)]
 pub enum OpType {
