+++ crates/vm/src/system/memory/offline.rs
@@ -1,4 +1,4 @@
-use std::{array, cmp::max, fmt::Debug, sync::Arc};
+use std::{array, cmp::max, sync::Arc};
 
 use openvm_circuit_primitives::{
     assert_less_than::AssertLtSubAir, var_range::VariableRangeCheckerChip,
@@ -9,8 +9,18 @@ use rustc_hash::{FxHashMap, FxHashSet};
 use crate::system::memory::{
     adapter::{AccessAdapterRecord, AccessAdapterRecordKind},
     offline_checker::{MemoryBridge, MemoryBus},
-    MemoryAuxColsFactory, TimestampedEquipartition, TimestampedValues,
+    online::Address,
+    MemoryAuxColsFactory, MemoryImage, RecordId, TimestampedEquipartition, TimestampedValues,
 };
+
+pub const INITIAL_TIMESTAMP: u32 = 0;
+
+#[derive(Clone, Copy, PartialEq, Eq, Debug)]
+struct BlockData {
+    pointer: u32,
+    size: usize,
+    timestamp: u32,
+}
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct MemoryRecord<T> {
@@ -21,80 +31,6 @@ pub struct MemoryRecord<T> {
     pub data: Vec<T>,
     /// None if a read.
     pub prev_data: Option<Vec<T>>,
-}
-
-impl<T: Copy> MemoryRecord<T> {
-    pub fn value(&self) -> T {
-        assert!(self.data.len() == 1);
-        self.data[0]
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct RecordId(usize);
-
-#[derive(Debug, Clone)]
-pub enum MemoryLogEntry<T> {
-    Read {
-        address_space: u32,
-        pointer: u32,
-        len: usize,
-    },
-    Write {
-        address_space: u32,
-        pointer: u32,
-        data: Vec<T>,
-    },
-    IncrementTimestampBy(u32),
-}
-
-/// Represents a single or batch memory write operation.
-/// Can be used to generate [MemoryWriteAuxCols].
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct MemoryWriteRecord<T, const N: usize> {
-    pub address_space: T,
-    pub pointer: T,
-    pub timestamp: u32,
-    pub prev_timestamp: u32,
-    pub data: [T; N],
-    pub prev_data: [T; N],
-}
-
-impl<T: Copy> MemoryWriteRecord<T, 1> {
-    pub fn value(&self) -> T {
-        self.data[0]
-    }
-}
-
-/// Represents a single or batch memory read operation.
-///
-/// Also used for "reads" from address space 0 (immediates).
-/// Can be used to generate [MemoryReadAuxCols] or [MemoryReadOrImmediateAuxCols].
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct MemoryReadRecord<T, const N: usize> {
-    pub address_space: T,
-    pub pointer: T,
-    pub timestamp: u32,
-    pub prev_timestamp: u32,
-    pub data: [T; N],
-}
-
-impl<T: Copy> MemoryReadRecord<T, 1> {
-    pub fn value(&self) -> T {
-        self.data[0]
-    }
-}
-
-pub const INITIAL_TIMESTAMP: u32 = 0;
-
-/// (address_space, pointer)
-pub(crate) type Address = (u32, u32);
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-struct BlockData {
-    pointer: u32,
-    size: usize,
-    timestamp: u32,
 }
 
 #[derive(Debug)]
@@ -504,111 +440,6 @@ impl<F: PrimeField32> OfflineMemory<F> {
     }
 }
 
-pub type MemoryImage<F> = FxHashMap<Address, F>;
-
-/// A simple data structure to read to/write from memory.
-///
-/// Stores a log of memory accesses to reconstruct aspects of memory state for trace generation.
-#[derive(Debug, Clone)]
-pub struct Memory<F> {
-    pub(super) data: FxHashMap<Address, F>,
-    pub(super) log: Vec<MemoryLogEntry<F>>,
-    timestamp: u32,
-}
-
-impl<F: PrimeField32> Default for Memory<F> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<F: PrimeField32> Memory<F> {
-    pub fn new() -> Self {
-        Self {
-            data: MemoryImage::default(),
-            timestamp: INITIAL_TIMESTAMP + 1,
-            log: vec![],
-        }
-    }
-
-    /// Instantiates a new `Memory` data structure from an image.
-    pub fn from_image(image: MemoryImage<F>) -> Self {
-        Self {
-            data: image,
-            timestamp: INITIAL_TIMESTAMP + 1,
-            log: vec![],
-        }
-    }
-
-    fn last_record_id(&self) -> RecordId {
-        RecordId(self.log.len() - 1)
-    }
-
-    /// Writes an array of values to the memory at the specified address space and start index.
-    ///
-    /// Returns the `RecordId` for the memory record and the previous data.
-    pub fn write<const N: usize>(
-        &mut self,
-        address_space: u32,
-        pointer: u32,
-        values: [F; N],
-    ) -> (RecordId, [F; N]) {
-        assert!(N.is_power_of_two());
-
-        let prev_data = array::from_fn(|i| {
-            self.data
-                .insert((address_space, pointer + i as u32), values[i])
-                .unwrap_or(F::ZERO)
-        });
-
-        self.log.push(MemoryLogEntry::Write {
-            address_space,
-            pointer,
-            data: values.to_vec(),
-        });
-        self.timestamp += 1;
-
-        (self.last_record_id(), prev_data)
-    }
-
-    /// Reads an array of values from the memory at the specified address space and start index.
-    pub fn read<const N: usize>(&mut self, address_space: u32, pointer: u32) -> (RecordId, [F; N]) {
-        assert!(N.is_power_of_two());
-
-        self.log.push(MemoryLogEntry::Read {
-            address_space,
-            pointer,
-            len: N,
-        });
-
-        let values = if address_space == 0 {
-            assert_eq!(N, 1, "cannot batch read from address space 0");
-            [F::from_canonical_u32(pointer); N]
-        } else {
-            self.range_array::<N>(address_space, pointer)
-        };
-        self.timestamp += 1;
-        (self.last_record_id(), values)
-    }
-
-    pub fn increment_timestamp_by(&mut self, amount: u32) {
-        self.timestamp += amount;
-        self.log.push(MemoryLogEntry::IncrementTimestampBy(amount))
-    }
-
-    pub fn timestamp(&self) -> u32 {
-        self.timestamp
-    }
-
-    pub fn get(&self, address_space: u32, pointer: u32) -> F {
-        *self.data.get(&(address_space, pointer)).unwrap_or(&F::ZERO)
-    }
-
-    fn range_array<const N: usize>(&self, address_space: u32, pointer: u32) -> [F; N] {
-        array::from_fn(|i| self.get(address_space, pointer + i as u32))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
@@ -617,12 +448,11 @@ mod tests {
     use openvm_stark_backend::p3_field::FieldAlgebra;
     use openvm_stark_sdk::p3_baby_bear::BabyBear;
 
-    use super::{BlockData, Memory};
+    use super::{BlockData, MemoryRecord, OfflineMemory};
     use crate::system::memory::{
         adapter::{AccessAdapterRecord, AccessAdapterRecordKind},
-        controller::MemoryRecord,
         offline_checker::MemoryBus,
-        MemoryImage, OfflineMemory, TimestampedValues,
+        MemoryImage, TimestampedValues,
     };
 
     macro_rules! bb {
@@ -696,18 +526,29 @@ mod tests {
 
     #[test]
     fn test_write_read_initial_block_len_1() {
-        let mut memory = Memory::new();
+        let initial_memory = MemoryImage::default();
+        let mut memory = OfflineMemory::<BabyBear>::new(
+            initial_memory,
+            1,
+            MemoryBus(0),
+            Arc::new(VariableRangeCheckerChip::new(VariableRangeCheckerBus::new(
+                1, 29,
+            ))),
+            29,
+        );
         let address_space = 1;
 
-        memory.write(address_space, 0, bba![1, 2, 3, 4]);
+        memory.write(address_space, 0, bbvec![1, 2, 3, 4]);
 
-        let (_, data) = memory.read::<2>(address_space, 0);
-        assert_eq!(data, bba![1, 2]);
+        memory.read(address_space, 0, 2);
+        let read_record = memory.last_record();
+        assert_eq!(read_record.data, bba![1, 2]);
 
-        memory.write(address_space, 2, bba![100]);
+        memory.write(address_space, 2, bbvec![100]);
 
-        let (_, data) = memory.read::<4>(address_space, 0);
-        assert_eq!(data, bba![1, 2, 100, 4]);
+        memory.read(address_space, 0, 4);
+        let read_record = memory.last_record();
+        assert_eq!(read_record.data, bba![1, 2, 100, 4]);
     }
 
     #[test]
